Modularize frontend: add Vite, extract script.js and style.css modules
Implements the agreed-upon recommendations from the Antigravity code review:

1. Add Vite as frontend bundler with build/preview npm scripts
2. Extract 5 ES modules from script.js (5020→4208 lines):
   - modules/helpers.js: utility functions (sleep, toast, confetti, etc.)
   - modules/session.js: session time tracking and heartbeat
   - modules/gamification.js: XP, levels, celebrations, leaderboard
   - modules/billing.js: usage gating and upgrade prompts
   - modules/audio.js: TTS playback queue with pause/resume/speed
3. Extract 4 CSS modules from style.css into css/base/:
   - variables.css, typography.css, buttons.css, landing.css
4. Create css/main.css as Vite CSS entry point

style.css and all 22 HTML pages are kept backward-compatible.
