+++ docs/antigravity-review-evaluation.md
@@ -62,3 +62,51 @@ This document evaluates the Antigravity review of the MathmatixAI codebase, veri
 3. Extract `style.css` into component/page CSS files
 4. Leave backend architecture as-is (already modular)
 5. Framework migration only after frontend is modularized and bundled
+
+---
+
+## Implementation Status
+
+The following recommendations have been implemented:
+
+### 1. Vite Bundler (DONE)
+- Installed Vite as dev dependency
+- Created `vite.config.js` with multi-entry configuration
+- Added `npm run build` and `npm run build:preview` scripts
+- CSS entry point: `public/css/main.css`
+
+### 2. script.js Modularization (DONE — Phase 1)
+Reduced `script.js` from **5,020 → 4,208 lines** by extracting 5 ES modules:
+
+| Module | Lines | What it contains |
+|--------|-------|------------------|
+| `modules/helpers.js` | 92 | sleep, getGraphColor, generateSpeakableText, showToast, escapeHtml, triggerConfetti |
+| `modules/session.js` | 100 | Session time tracking, heartbeat, visibility detection |
+| `modules/gamification.js` | 259 | XP animation, level-up celebrations, leaderboard, quests, tutor unlock |
+| `modules/billing.js` | 132 | Billing status, usage gating, upgrade prompts, Stripe checkout |
+| `modules/audio.js` | 304 | TTS playback queue, pause/resume/speed, Web Audio API |
+
+**Total extracted: 887 lines across 5 focused modules**
+
+Remaining in script.js (candidates for future extraction):
+- Chat UI rendering (appendMessage, streaming, suggestions)
+- Message queue (send, queue, process)
+- IEP accommodations
+- Assessment system
+- Whiteboard
+- Event listeners
+
+### 3. style.css Modularization (DONE — Phase 1)
+Extracted 4 CSS modules from `style.css` (6,696 lines):
+
+| Module | Lines | What it contains |
+|--------|-------|------------------|
+| `css/base/variables.css` | 165 | CSS custom properties, reset, base styles |
+| `css/base/typography.css` | 112 | Typography, form elements |
+| `css/base/buttons.css` | 140 | All button variants and card styles |
+| `css/base/landing.css` | 417 | Landing page sections (hero, pricing, FAQ, etc.) |
+
+**Total extracted: 834 lines across 4 focused modules**
+
+`style.css` is kept intact for backward compatibility with all 22 HTML pages.
+The `css/main.css` entry point imports the modular files for Vite production builds.
