+++ package.json
@@ -6,6 +6,8 @@
   "scripts": {
     "start": "node server.js",
     "dev": "nodemon server.js",
+    "build": "vite build",
+    "build:preview": "vite preview",
     "test": "jest --coverage --verbose",
     "test:watch": "jest --watch",
     "test:unit": "jest --testPathPattern=tests/unit",
@@ -49,6 +51,7 @@
     "cors": "^2.8.5",
     "csv-parser": "^3.2.0",
     "dompurify": "^3.3.1",
+    "dotenv": "^17.2.4",
     "elevenlabs": "^1.59.0",
     "express": "^4.22.1",
     "express-rate-limit": "^7.5.1",
@@ -76,12 +79,12 @@
     "stripe": "^20.3.1",
     "uuid": "^9.0.1",
     "winston": "^3.11.0",
-    "winston-daily-rotate-file": "^4.7.1",
-    "dotenv": "^17.2.4"
+    "winston-daily-rotate-file": "^4.7.1"
   },
   "devDependencies": {
     "jest": "^29.7.0",
     "nodemon": "^3.1.11",
-    "supertest": "^7.0.0"
+    "supertest": "^7.0.0",
+    "vite": "^7.3.1"
   }
 }
