+++ public/css/base/variables.css
@@ -1,0 +1,165 @@
+@charset "UTF-8";
+/* style.css — PHASE 2: Frontend HTML Structure & CSS Styling - Batch 1 (CORRECTED & MERGED) */
+
+/* --- 1. Root Variables & Basic Reset --- */
+
+:root {
+    /* Sizing & Layout */
+    --header-height: 72px;
+    --footer-height: 48px;
+    --page-horizontal-padding: 20px;
+    --section-gap: 48px;
+    --content-max-width: 1100px;
+
+    /* Size Controls */
+    --logo-header-height: 44px;
+    --icon-size-default: 2.5rem;
+
+    /* Spacing Scale (4px base) */
+    --space-1: 4px;
+    --space-2: 8px;
+    --space-3: 12px;
+    --space-4: 16px;
+    --space-5: 20px;
+    --space-6: 24px;
+    --space-8: 32px;
+    --space-10: 40px;
+    --space-12: 48px;
+    --space-16: 64px;
+
+    /* Dashboard Specific Sizing */
+    --dashboard-panel-padding: 20px;
+    --dashboard-panel-header-height: 40px;
+    --dashboard-panel-header-margin-bottom: 15px;
+    --sidebar-width-desktop: 280px;
+    --chat-min-width: 400px;
+
+    /* Brand Colors */
+    --clr-primary-teal: #12B3B3;
+    --clr-primary-teal-light: #E6F7F7;
+    --clr-primary-teal-dark: #0E8F8F;
+    --clr-accent-hotpink: #FF3B7F;
+    --clr-accent-hotpink-light: #FFF0F5;
+    --clr-success-green: #16C86D;
+    --clr-success-green-light: #E8FAF0;
+    --clr-warning-gold: #FFC24B;
+    --clr-warning-gold-light: #FFF8EB;
+    --clr-error-red: #FF4E4E;
+    --clr-error-red-light: #FFF0F0;
+    --clr-bg-light: #FFFFFF;
+    --clr-bg-subtle: #F8FAFB;
+    --clr-bg-dark: #0F1A24;
+    --clr-text: #18202B;
+    --clr-text-dim: #5B6876;
+    --clr-border: #E2E8F0;
+    --clr-border-light: #F1F5F9;
+
+    /* RGB versions for rgba() usage */
+    --clr-primary-teal-rgb: 18, 179, 179;
+    --clr-accent-hotpink-rgb: 255, 59, 127;
+    --clr-success-green-rgb: 22, 200, 109;
+
+    /* Text Colors */
+    --text-primary: var(--clr-text);
+    --text-secondary: var(--clr-text-dim);
+    --text-light: var(--clr-bg-light);
+    --text-dark: var(--clr-text);
+    --text-green: var(--clr-success-green);
+    --text-hotpink: var(--clr-accent-hotpink);
+
+    /* Shadows — Multi-layer for richer depth */
+    --shadow-xs: 0 1px 2px rgba(0,0,0,0.04);
+    --shadow-sm: 0 1px 3px rgba(0,0,0,0.06), 0 1px 2px rgba(0,0,0,0.04);
+    --shadow-md: 0 4px 6px -1px rgba(0,0,0,0.08), 0 2px 4px -1px rgba(0,0,0,0.04);
+    --shadow-lg: 0 10px 15px -3px rgba(0,0,0,0.08), 0 4px 6px -2px rgba(0,0,0,0.04);
+    --shadow-xl: 0 20px 25px -5px rgba(0,0,0,0.08), 0 10px 10px -5px rgba(0,0,0,0.03);
+    --shadow-card-primary: 0 4px 14px rgba(var(--clr-primary-teal-rgb), 0.12);
+    --shadow-card-hover: 0 12px 28px rgba(var(--clr-primary-teal-rgb), 0.16);
+    --shadow-glow-teal: 0 0 20px rgba(var(--clr-primary-teal-rgb), 0.15);
+    --shadow-glow-pink: 0 0 20px rgba(var(--clr-accent-hotpink-rgb), 0.15);
+
+    /* Typography Defaults */
+    --font-family-headlines: "Inter", -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+    --font-family-body: "Inter", -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+    --font-family-code-math: "ui-monospace", "SF Mono", Menlo, Consolas, monospace;
+    --font-weight-headlines: 700;
+    --font-weight-semibold: 600;
+    --font-weight-medium: 500;
+    --font-weight-body: 400;
+    --brand-base-font-size: 1rem;
+    --brand-h1-font-size: clamp(2rem, 5vw, 2.8rem);
+    --brand-h2-font-size: clamp(1.6rem, 4vw, 2.2rem);
+    --brand-h3-font-size: clamp(1.15rem, 3vw, 1.4rem);
+    --line-height-tight: 1.2;
+    --line-height-normal: 1.5;
+    --line-height-relaxed: 1.65;
+    --letter-spacing-tight: -0.02em;
+    --letter-spacing-normal: 0;
+
+    /* Border Radius */
+    --radius-sm: 6px;
+    --radius-base: 10px;
+    --radius-md: 12px;
+    --radius-lg: 16px;
+    --radius-xl: 20px;
+    --radius-full: 9999px;
+    --border-radius-base: 10px;
+
+    /* Transitions */
+    --transition-fast: 150ms cubic-bezier(0.4, 0, 0.2, 1);
+    --transition-base: 200ms cubic-bezier(0.4, 0, 0.2, 1);
+    --transition-smooth: 300ms cubic-bezier(0.22, 1, 0.36, 1);
+    --transition-spring: 400ms cubic-bezier(0.34, 1.56, 0.64, 1);
+
+    /* Z-index scale */
+    --z-base: 1;
+    --z-dropdown: 100;
+    --z-sticky: 200;
+    --z-header: 1000;
+    --z-sidebar: 1100;
+    --z-drawer: 2000;
+    --z-modal: 9999;
+    --z-toast: 10000;
+}
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  font-family: var(--font-family-body);
+}
+
+/* --- Core Layout for Fixed Header/Footer & Scrollable Content --- */
+
+html {
+    scroll-behavior: smooth;
+    -webkit-text-size-adjust: 100%;
+    text-size-adjust: 100%;
+}
+
+body {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+    min-height: 100dvh; /* Dynamic viewport height for mobile */
+    background-color: var(--clr-bg-light);
+    color: var(--text-primary);
+    font-size: var(--brand-base-font-size);
+    line-height: var(--line-height-relaxed);
+    position: relative;
+    min-width: 320px;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    text-rendering: optimizeLegibility;
+    overflow-x: hidden;
+}
+
+/* Respect reduced motion preferences */
+@media (prefers-reduced-motion: reduce) {
+    *, *::before, *::after {
+        animation-duration: 0.01ms !important;
+        animation-iteration-count: 1 !important;
+        transition-duration: 0.01ms !important;
+        scroll-behavior: auto !important;
+    }
+}
