+++ vite.config.js
@@ -1,0 +1,29 @@
+import { defineConfig } from 'vite';
+import { resolve, dirname } from 'path';
+import { fileURLToPath } from 'url';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+export default defineConfig({
+  root: 'public',
+  build: {
+    outDir: '../dist',
+    emptyOutDir: true,
+    rollupOptions: {
+      input: {
+        chat: resolve(__dirname, 'public/js/script.js'),
+        style: resolve(__dirname, 'public/css/main.css'),
+      },
+      output: {
+        entryFileNames: 'js/[name].[hash].js',
+        chunkFileNames: 'js/chunks/[name].[hash].js',
+        assetFileNames: 'assets/[name].[hash][extname]',
+      },
+    },
+  },
+  resolve: {
+    alias: {
+      '@modules': resolve(__dirname, 'public/js/modules'),
+    },
+  },
+});
